# docs/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "Localhost",
    "url" : "http://localhost:3000"
  } ],
  "info" : {
    "description" : "An emergency response and contact management API.",
    "version" : "1.0.0",
    "title" : "Bmore Responsive",
    "contact" : {
      "email" : "hello@codeforbaltimore.org"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags" : [ {
    "name" : "developer",
    "description" : "Operations available to regular developers"
  }, {
    "name" : "user",
    "description" : "Operations related to system users"
  }, {
    "name" : "userRole",
    "description" : "Operations related to user roles"
  } ],
  "paths" : {
    "/health" : {
      "get" : {
        "tags" : [ "developer" ],
        "summary" : "returns a healthcheck",
        "description" : "This returns simple system information\n",
        "responses" : {
          "200" : {
            "description" : "the system healthcheck",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HealthcheckItem"
                }
              }
            }
          },
          "500" : {
            "description" : "System error"
          }
        }
      }
    },
    "/user/login" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "logs in a system user",
        "description" : "The system login endpoint.",
        "requestBody" : {
          "description" : "The body of the payload",
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "email", "password" ],
                "properties" : {
                  "email" : {
                    "type" : "string",
                    "format" : "email",
                    "example" : "test@test.test"
                  },
                  "password" : {
                    "type" : "string",
                    "example" : "test"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "login token"
          },
          "400" : {
            "description" : "Invalid credentials"
          }
        }
      }
    },
    "/user" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "creates a new user",
        "description" : "The new user endpoint.",
        "requestBody" : {
          "description" : "The body of the payload",
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "email", "password" ],
                "properties" : {
                  "email" : {
                    "type" : "string",
                    "format" : "email",
                    "example" : "test@test.test"
                  },
                  "password" : {
                    "type" : "string",
                    "format" : "password",
                    "example" : "test"
                  },
                  "roles" : {
                    "type" : "array",
                    "example" : [1]
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "user created"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      },
      "get" : {
        "tags" : [ "user" ],
        "summary" : "returns all system users",
        "description" : "By passing in the appropriate options, you can search for\navailable users in the system\n",
        "parameters" : [ {
          "in" : "header",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1ODA3NTM0MDUsImV4cCI6MTU4MDgzOTgwNX0.Q6W7Vo6By35yjZBeLKkk96s8LyqIE2G39AG1H3LRD9M"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserItem"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        }
      },
      "put" : {
        "tags" : [ "user" ],
        "summary" : "updates any single system user",
        "description" : "By sending a valid payload you can update a user.",
        "parameters" : [ {
          "in" : "header",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1ODA3NTM0MDUsImV4cCI6MTU4MDgzOTgwNX0.Q6W7Vo6By35yjZBeLKkk96s8LyqIE2G39AG1H3LRD9M"
          }
        } ],
        "requestBody" : {
          "description" : "The body of the payload.",
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "email", "password" ],
                "properties" : {
                  "email" : {
                    "type" : "string",
                    "format" : "email",
                    "example" : "test@test.test"
                  },
                  "password" : {
                    "type" : "string",
                    "format" : "password",
                    "example" : "test"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "user updated"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        }
      }
    },
    "/user/{email}" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "returns a single system user",
        "description" : "By passing the username, you can lookup a user.",
        "parameters" : [ {
          "in" : "path",
          "name" : "email",
          "schema" : {
            "type" : "string",
            "format" : "email"
          },
          "required" : true,
          "description" : "email of the user"
        }, {
          "in" : "header",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1ODA3NTM0MDUsImV4cCI6MTU4MDgzOTgwNX0.Q6W7Vo6By35yjZBeLKkk96s8LyqIE2G39AG1H3LRD9M"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "the user object",
            "content" : {
              "session/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserItem"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        }
      },
      "delete" : {
        "tags" : [ "user" ],
        "summary" : "deletes a single system user",
        "description" : "By sending a valid payload you can delete a user.",
        "parameters" : [ {
          "in" : "path",
          "name" : "email",
          "schema" : {
            "type" : "string",
            "format" : "email"
          },
          "required" : true,
          "description" : "the email of the user to delete."
        }, {
          "in" : "header",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1ODA3NTM0MDUsImV4cCI6MTU4MDgzOTgwNX0.Q6W7Vo6By35yjZBeLKkk96s8LyqIE2G39AG1H3LRD9M"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "user deleted"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        }
      }
    },
    "/userRole" : {
      "post" : {
        "tags" : [ "userRole" ],
        "summary" : "creates a new user role",
        "description" : "The new user role endpoint.",
        "parameters" : [ {
          "in" : "header",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1ODA3NTM0MDUsImV4cCI6MTU4MDgzOTgwNX0.Q6W7Vo6By35yjZBeLKkk96s8LyqIE2G39AG1H3LRD9M"
          }
        } ],
        "requestBody" : {
          "description" : "The body of the payload",
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "role" ],
                "properties" : {
                  "role" : {
                    "type" : "string",
                    "example" : "test"
                  },
                  "description" : {
                    "type" : "string",
                    "example" : "This is a test"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "user role created"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        }
      },
      "get" : {
        "tags" : [ "userRole" ],
        "summary" : "returns all system user roles",
        "description" : "By passing in the appropriate options, you can search for\navailable user roles in the system\n",
        "parameters" : [ {
          "in" : "header",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1ODA3NTM0MDUsImV4cCI6MTU4MDgzOTgwNX0.Q6W7Vo6By35yjZBeLKkk96s8LyqIE2G39AG1H3LRD9M"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserRoleItem"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        }
      },
      "put" : {
        "tags" : [ "userRole" ],
        "summary" : "updates any single system user role",
        "description" : "By sending a valid payload you can update a user role.",
        "parameters" : [ {
          "in" : "header",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1ODA3NTM0MDUsImV4cCI6MTU4MDgzOTgwNX0.Q6W7Vo6By35yjZBeLKkk96s8LyqIE2G39AG1H3LRD9M"
          }
        } ],
        "requestBody" : {
          "description" : "The body of the payload.",
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "id", "role" ],
                "properties" : {
                  "id" : {
                    "type" : "number",
                    "example" : 1
                  },
                  "role" : {
                    "type" : "string",
                    "example" : "test"
                  },
                  "description" : {
                    "type" : "string",
                    "example" : "This is a test 2"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "user updated"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        }
      }
    },
    "/userRole/{role}" : {
      "get" : {
        "tags" : [ "userRole" ],
        "summary" : "returns a single system user",
        "description" : "By passing the username, you can lookup a user.",
        "parameters" : [ {
          "in" : "path",
          "name" : "role",
          "schema" : {
            "type" : "string"
          },
          "required" : true,
          "description" : "the role to get"
        }, {
          "in" : "header",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1ODA3NTM0MDUsImV4cCI6MTU4MDgzOTgwNX0.Q6W7Vo6By35yjZBeLKkk96s8LyqIE2G39AG1H3LRD9M"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "the user object",
            "content" : {
              "session/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserRoleItem"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        }
      },
      "delete" : {
        "tags" : [ "userRole" ],
        "summary" : "deletes a single system user role",
        "description" : "By sending a valid payload you can delete a user role.",
        "parameters" : [ {
          "in" : "path",
          "name" : "role",
          "schema" : {
            "type" : "string"
          },
          "required" : true,
          "description" : "the user role to delete."
        }, {
          "in" : "header",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1ODA3NTM0MDUsImV4cCI6MTU4MDgzOTgwNX0.Q6W7Vo6By35yjZBeLKkk96s8LyqIE2G39AG1H3LRD9M"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "user deleted"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "HealthcheckItem" : {
        "type" : "object",
        "required" : [ "uptime", "environment", "version", "requestId" ],
        "properties" : {
          "uptime" : {
            "type" : "string",
            "example" : "42:42:42"
          },
          "environment" : {
            "type" : "string",
            "example" : "dev"
          },
          "version" : {
            "type" : "string",
            "example" : "0.0.1"
          },
          "requestId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "602e1bd9-3eb1-4a26-a0d1-6bf08e8c0a66"
          }
        }
      },
      "UserItem" : {
        "type" : "object",
        "required" : [ "email" ],
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "example" : "test@test.test"
          },
          "roles": {
            "type" : "array",
            "example" : [{
              "id" : 1,
              "role" : "Test",
              "description" : "This is a test."
            }]
          },
          "displayNamme" : {
            "type" : "string",
            "example" : "test"
          },
          "phone" : {
            "type" : "string",
            "example" : "123456789"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2020-01-21T13:45:52.348Z"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2020-01-21T13:45:52.348Z"
          }
        }
      },
      "UserRoleItem" : {
        "type" : "object",
        "required" : [ "id", "role" ],
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 1
          },
          "role" : {
            "type" : "string",
            "example" : "test"
          },
          "description" : {
            "type" : "string",
            "example" : "This is a test"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2020-01-21T13:45:52.348Z"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2020-01-21T13:45:52.348Z"
          }
        }
      }
    }
  }
}